requirements.txt
beautifulsoup4==4.5.1
requests==2.12.4
py-postgresql==1.1.0
sh==1.9